requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==6.2.2  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
redis==3.4.1 # https://github.com/andymccurdy/redis-py

# Django
# ------------------------------------------------------------------------------
django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.11.0  # https://github.com/niwinz/django-redis
django-extensions==2.2.9
# Wagtail
# ------------------------------------------------------------------------------
wagtail==2.8.1
wagtailfontawesome==1.2.1
-e git://github.com/aquilesC/wagtail-markdown.git@develop#egg=wagtail-markdown
# Newsletter
# ------------------------------------------------------------------------------
django-newsletter==0.9

# Comments
# ------------------------------------------------------------------------------
django-comments-xtd==2.5.0